requirements.txt
django 
django-cors-headers
torch
torchvision
wikipedia-api
rich
Pillow
axios
react-router-dom
djangorestframework
wikipedia-api
react-dom
djangorestframework-simplejwt